requirements.txt
crewai>=0.86.0
crewai-tools>=0.17.0
anthropic>=0.39.0
python-dotenv>=1.0.1
pydantic>=2.10.0
